+++ packages/sdk/src/config/type.ts
@@ -27,6 +27,11 @@ export type FactoriesAddresses = {
 	erc1155: Address
 }
 
+export type CryptoPunksConfig = {
+	marketContract: Address
+	wrapperContract: Address
+}
+
 export type EthereumConfig = {
 	basePath: string
 	chainId: number
@@ -36,5 +41,6 @@ export type EthereumConfig = {
 	openSea: OpenSeaConfig
 	factories: FactoriesAddresses
 	weth: Address
-	auction: Address
+	auction: Address,
+	cryptoPunks: CryptoPunksConfig
 }